FE và sửa BE lại

+++ order-service/src/main/java/com/mutrapro/order_service/repository/OrderRepository.java
@@ -6,5 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.mutrapro.order_service.model.Order;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-}
+public interface OrderRepository extends JpaRepository<Order, Long> {}
